refactoring finished styling not finished
